First steps of actual QJniObject templating
This also changes NativeProxyGenerator a bit:
1) There is now an additional header_final_postamble template used. This
allows generating a header guard.
2) The handling of jte exceptions no longer swallows the exception, its data is now printed.

Other than that, we add a new android sample, which uses the QJniObject templates.
The first part of that is adding a data member, and the <QJniObject> include, no other
changes yet, other than the use of an include guard instead of a #pragma once.