Make the JTE builds use different directories for debug/release builds
This fixes our longest-standing problem; when run in parallel, the build
sometimes (often) fails, because parallel builds cause JTE to write
to the same java class files concurrently.

The problem is cured by defining a task-specific property that is then passed
down to Jenny as the directory suffix to use for the JTE output directory.

Change-Id: I1f49a3af7f8d1102dc06cc1705c0ff1868477568
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>